[Qwen3] StateDictAdapter support for MoE model
